+++ src/main/kotlin/cloud/glitchdev/rfu/achievement/achievements/hotspot/BuffCollectorAchievement.kt
@@ -2,6 +2,7 @@ package cloud.glitchdev.rfu.achievement.achievements.hotspot
 
 import cloud.glitchdev.rfu.achievement.*
 import cloud.glitchdev.rfu.achievement.types.NumericAchievement
+import cloud.glitchdev.rfu.data.fishing.Bait
 import cloud.glitchdev.rfu.events.managers.SeaCreatureCatchEvents.registerSeaCreatureCatchEvent
 
 @Achievement
@@ -24,7 +25,7 @@ object BuffCollectorAchievement : NumericAchievement() {
     )
 
     override fun setupListeners() {
-        activeListeners.add(registerSeaCreatureCatchEvent { _, _, hotspot, _ ->
+        activeListeners.add(registerSeaCreatureCatchEvent { _, _, hotspot, _, _ ->
             if (hotspot != null) {
                 val buff = hotspot.buff