+++ src/main/kotlin/cloud/glitchdev/rfu/feature/mob/SeaCreatureHour.kt
@@ -30,7 +30,7 @@ object SeaCreatureHour : Feature {
     var total: Int = 0
 
     override fun onInitialize() {
-        registerSeaCreatureCatchEvent { _, isDoubleHook, _, _ ->
+        registerSeaCreatureCatchEvent { _, isDoubleHook, _, _, _ ->
             handleCatch(isDoubleHook)
             RareSCDisplay.updateState()
         }